fix offsets in mdf4 structures with non-byte members
When generating signals for a structure, the designated size in bytes of
an element is continuously accumulated to calculate the offsets of
all elements. When an element does not have a size that is a multiple
of 8 (i.e., a boolean flag), the byte size was calculated to be zero,
which means that the subsequent channel erroneously received the same
offset. Fix by always assuming at least 1 byte for the size of an
element.